+++ libs/docs/platform/icon-tab-bar/platform-icon-tab-bar-docs.component.ts
@@ -19,6 +19,7 @@ import { PlatformIconTabBarIconTypeExampleComponent } from './examples/platform-
 import { PlatformIconTabBarProcessTypeExampleComponent } from './examples/platform-icon-tab-bar-process-type-example/platform-icon-tab-bar-process-type-example.component';
 import { PlatformIconTabBarTextTypeExampleComponent } from './examples/platform-icon-tab-bar-text-type-example/platform-icon-tab-bar-text-type-example.component';
 import { IconTabBarTabsContentExampleComponent } from './examples/tabs-content/icon-tab-bar-tabs-content-example.component';
+import { IconTabBarTwoClickAreaExampleComponent } from './examples/two-click-area/icon-tab-bar-two-click-area-example.component';
 
 const iconTabBarTextTypeHtml =
     'platform-icon-tab-bar-text-type-example/platform-icon-tab-bar-text-type-example.component.html';
@@ -56,6 +57,9 @@ const iconTabBarClosableTs = 'closable-tabs/icon-tab-bar-closable-tabs-example.c
 const iconTabBarProjectedHtml = 'tabs-content/icon-tab-bar-tabs-content-example.component.html';
 const iconTabBarProjectedTs = 'tabs-content/icon-tab-bar-tabs-content-example.component.ts';
 
+const iconTabBarTwoClickHtml = 'two-click-area/icon-tab-bar-two-click-area-example.component.html';
+const iconTabBarTwoClickTs = 'two-click-area/icon-tab-bar-two-click-area-example.component.ts';
+
 const iconTypeConfigs = 'config-for-examples/icon-type-config.ts';
 
 const textTypeConfigs = 'config-for-examples/text-type.config.ts';
@@ -89,7 +93,8 @@ export interface IconTabBarItem extends TabConfig {
         MessageStripComponent,
         PlatformIconTabBarConfigurablePaddingsExampleComponent,
         IconTabBarClosableTabsExampleComponent,
-        IconTabBarTabsContentExampleComponent
+        IconTabBarTabsContentExampleComponent,
+        IconTabBarTwoClickAreaExampleComponent
     ]
 })
 export class PlatformIconTabBarDocsComponent {
@@ -267,6 +272,29 @@ export class PlatformIconTabBarDocsComponent {
             fileName: 'icon-tab-bar-tabs-content-example',
             component: 'IconTabBarTabsContentExampleComponent',
             path: 'example'
+        }
+    ];
+
+    iconTabBarTwoClickFiles: ExampleFile[] = [
+        {
+            language: 'html',
+            code: getAssetFromModuleAssets(iconTabBarTwoClickHtml),
+            fileName: 'icon-tab-bar-two-click-area-example',
+            path: 'example'
+        },
+        {
+            language: 'typescript',
+            code: getAssetFromModuleAssets(iconTabBarTwoClickTs),
+            fileName: 'icon-tab-bar-two-click-area-example',
+            component: 'IconTabBarTwoClickAreaExampleComponent',
+            path: 'example'
+        },
+        {
+            language: 'typescript',
+            code: getAssetFromModuleAssets(textTypeConfigs),
+            fileName: 'text-type.config',
+            path: 'config-for-examples',
+            pure: true
         }
     ];
 }