+++ libs/i18n/src/lib/models/fd-language-key.ts
@@ -1,0 +1,7 @@
+import { Nullable } from '@fundamental-ngx/cdk/utils';
+
+export type FdLanguageKeyArgs = Nullable<Record<string, string | number | boolean>>;
+
+export type FdLanguageKeyFunction<T> = T extends undefined ? () => string : (args: T) => string;
+
+export type FdLanguageKey<T = undefined> = string | FdLanguageKeyFunction<T>;