+++ libs/i18n/src/lib/models/patch-language.ts
@@ -1,3 +1,3 @@
-import { FdLanguage } from './lang';
+import { FdLanguage } from './fd-language';
 
 export type FdLanguagePatch = { [K in keyof FdLanguage]?: Partial<FdLanguage[K]> };